backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.9
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
flake8==4.0.1
fonttools==4.28.2
fpdf==1.7.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
setuptools-scm==6.3.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
